# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo swagger",
        "version": "any"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "users list"
                ],
                "summary": "Get list users",
                "description": "Returns list users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/users/create": {
            "post": {
                "tags": [
                    "user create"
                ],
                "description": "create a user",
                "operationId": "create user",
                "requestBody": {
                    "description": "Fill into formdata",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "nam"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "nam@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "number",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Created!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/show/{id}": {
            "get": {
                "tags": [
                    "user show detail"
                ],
                "summary": "show user",
                "description": "show an user",
                "operationId": "show user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/update/{id}": {
            "post": {
                "tags": [
                    "user update"
                ],
                "summary": "update user by id",
                "description": "update an user",
                "operationId": "update an user by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fill into formdata",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "nam"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "nam@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "number",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/delete/{id}": {
            "delete": {
                "tags": [
                    "user deletion"
                ],
                "summary": "delete user",
                "description": "delete an user",
                "operationId": "delete user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}